Numbers fixed

+++ scenedsl.h
@@ -42,7 +42,7 @@ struct _lo {
 #define SC_MIX(A, B, C) DF_MIX, A, B, C
 #define SC_MIN(A, B) DF_MIN, A, B
 #define SC_MAX(A, B) DF_MAX, A, B
-#define SC_FIXED(f) (GF_NUMBER | _hi<(int)((f) * 256)>::v), _lo<(int)((f) * 256)>::v
+#define SC_FIXED(f) (uint8_t)(GF_NUMBER | _hi<(int)((f) * 256)>::v), _lo<(int)((f) * 256)>::v
 #define SC_TIME(n) GF_TIME, n
 #define SC_TIME2(n) GF_TIME2, n
 #define SC_CLAMP(gfarg, gfmin, gfmax) GF_CLAMP, gfarg, gfmin, gfmax
